Revert "trying to merge json"
This reverts commit 5556e8654bfefc75c704fc49ec8eb0a46fe6b88a, reversing
changes made to f1d8edb21aeee4f0506e232e007b60201fc3d6e4.

+++ Falcon/Falcon/Rendering/BufferDefinitions/VertexArray.h
@@ -19,7 +19,7 @@ public:
 	void Bind() const;
 	void Unbind() const;
 
-	void AddVertexAttribPointer(u32 loc, u32 size, u32 type, u32 normalize, u32 byteSize, u64 offset, u32 divisor);
+	void AddVertexAttribPointer(u32 loc, u32 size, u32 type, u32 normalize, u32 byteSize, u32 offset, u32 divisor);
 };
 
 #endif // !1
